% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Validation_Data.R
\name{build_valid_fixed_covs}
\alias{build_valid_fixed_covs}
\title{Generate fixed covariates for validation dataset (i.e. intercept, X and C)}
\usage{
build_valid_fixed_covs(K, N)
}
\arguments{
\item{K}{Number of groups}

\item{N}{Number of observations per group}
}
\value{
A list of tibbles. One entry for each group, containing all observations for that group.
}
\description{
Generate fixed covariates for validation dataset (i.e. intercept, X and C)
}
\examples{
K=2
N=10
build_valid_fixed_covs(K, N)
}
