% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Bootstrap.R
\name{check_bootstrap_inputs}
\alias{check_bootstrap_inputs}
\title{Check whether enough information has been provided to fit the requested flavour of bootstrap}
\usage{
check_bootstrap_inputs(
  data = NULL,
  mod_Y = NULL,
  mod_M = NULL,
  parametric = TRUE
)
}
\arguments{
\item{data}{Original dataset. Optional for parametric (as long as \code{mod_Y} and \code{mod_M} are supplied), required for non-parametric, .}

\item{mod_Y}{Regression model for predicting \code{Y}. Recommended but optional for parametric (as long as \code{data} is supplied), not used for non-parametric.}

\item{mod_M}{Regression model for predicting \code{M}. Recommended but optional for parametric (as long as \code{data} is supplied), not used for non-parametric.}

\item{parametric}{Indicates whether to perform parametric or non-parametric bootstrap.}
}
\value{
\code{TRUE} if there is sufficient information available to fit the requested bootstrap. \code{FALSE} otherwise. See arguments for restrictions.
}
\description{
Check whether enough information has been provided to fit the requested flavour of bootstrap
}
\examples{
data = make_validation_data()
mod_Y = fit_mod_Y(data)
mod_M = fit_mod_M(data)

# These settings are fine
check_bootstrap_inputs(mod_Y = mod_Y, mod_M = mod_M, parametric = TRUE)
check_bootstrap_inputs(data = data, parametric = FALSE)

# This setting is not supported
tryCatch({
    check_bootstrap_inputs(mod_Y = mod_Y, mod_M = mod_M, parametric = FALSE)
    print("No error found. This represents a failure of check_bootstrap_inputs()")
  },
  error = function(e) e
)


# This setting is okay, but probably inefficient
check_bootstrap_inputs(data = data, parametric = TRUE)
}
