% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_at}
\alias{delete_at}
\title{Remove the elements of X corresponding to the indices in inds}
\usage{
delete_at(X, inds)
}
\arguments{
\item{X}{A vector}

\item{inds}{A vector of indices}
}
\value{
A copy of \code{X} with the elements given in \code{inds} removed.
}
\description{
This function is a wrapper for negative indexing, with syntax inspired by the \code{Julia} function \verb{deleteat!()}. If any of the indices in \code{inds} are larger than the length of \code{X}, throw an error.
}
\examples{
A = c(1, 3, 5, 7)
delete_at(A, c(2,3))
}
\keyword{internal}
