% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Regression_Models.R
\name{fit_mod_M}
\alias{fit_mod_M}
\title{Fit a regression model to the observed data with mediator as response variable}
\usage{
fit_mod_M(data)
}
\arguments{
\item{data}{The observed data. A data.frame-like object (e.g. a tibble).}
}
\value{
A GLMM fit using \code{glmer} from the \code{lme4} package.
}
\description{
Fit a regression model to the observed data with mediator as response variable
}
\examples{
data = 0
fit_mod_M(data)
}
