% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Regression_Models.R
\name{fit_mod_M_formal}
\alias{fit_mod_M_formal}
\title{Fit regression model for mediator (\code{M}), assuming that data have the required structure}
\usage{
fit_mod_M_formal(data_formal)
}
\arguments{
\item{data_formal}{A formal data frame. See documentation for requirements.}
}
\value{
A GLMM fit using \code{glmer} from the \code{lme4} package.
}
\description{
Fit regression model for mediator (\code{M}), assuming that data have the required structure
}
\details{
I make strong assumptions about the structure of the data frame which is passed to this function. Such a structured data frame is referred to as \code{data_formal}. The requirements are as follows:
\itemize{
\item Must contain columns named Y, M, X and group
\item May also contain other columns. Each such extra column must be a confounder.      !!!! I'm not sure yet whether I want to require that these columns be named C1, C2, ...#'
}
}
\examples{
n = 20
K = 3
all_reg_pars = make_all_reg_pars()
data_formal = make_validation_data(n, K, all_reg_pars)

fit_mod_M_formal(data_formal)
}
