% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Regression_Models.R
\name{fit_mod_Y}
\alias{fit_mod_Y}
\title{Fit a regression model to the observed data with outcome as response variable}
\usage{
fit_mod_Y(data)
}
\arguments{
\item{data}{The observed data. A data.frame.}
}
\value{
A GLMM fit using \code{glmer} from the \code{lme4} package.
}
\description{
Fit a regression model to the observed data with outcome as response variable
}
\examples{
n = 20
K = 3
all_reg_pars = make_all_reg_pars()
data = make_validation_data(n, K, all_reg_pars)

fit_mod_Y(data)
}
