% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Mixed_Effects.R
\name{fixed_plus_mixed}
\alias{fixed_plus_mixed}
\title{Augment vector of fixed effects by adding random effects to get mixed}
\usage{
fixed_plus_mixed(fix_effs, ran_effs_by_K, mix_eff_ind_pairs)
}
\arguments{
\item{fix_effs}{Vector of fixed effects}

\item{ran_effs_by_K}{List containing one vector of random effects for each group}

\item{mix_eff_ind_pairs}{List of pairs of indices, (i,j), indicating which fixed (i) and random (j) effects have the same covariate. Note: Don't forget about the intercept if necessary}
}
\value{
A list containing one vector of fixed and mixed effects for each group.
}
\description{
Augment vector of fixed effects by adding random effects to get mixed
}
\examples{
fix_effs = c(1,2,3)
K = 2
ran_effs_by_K = list(1, 2)
mix_eff_ind_pairs = list(c(2,1))
fixed_plus_mixed(fix_effs, ran_effs_by_K, mix_eff_ind_pairs)
}
