% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_index}
\alias{get_index}
\title{Extract the ith element of each vector in all_vecs}
\usage{
get_index(all_vecs, i)
}
\arguments{
\item{all_vecs}{A list of vectors. Throws an error if any contain fewer than i entries.}

\item{i}{Index to extract.}
}
\value{
A vector of the same length as all_vecs, containing the ith component of each element in all_vecs. If any components of all_vecs have fewer than i elements, get_index() will throw an error.
}
\description{
Extract the ith element of each vector in all_vecs
}
\examples{
A = c(1,2,3)
B = c("a", "b")
C = 1:1000
all_vecs = list(A, B, C)
get_index(all_vecs, 2)
}
