% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_index_list}
\alias{get_index_list}
\title{Extract the \code{i}th element of each list in \code{all_lists}}
\usage{
get_index_list(all_lists, i)
}
\arguments{
\item{all_lists}{A list of lists. Throws an error if any contain fewer than \code{i} entries.}

\item{i}{Index to extract.}
}
\value{
A list of the same length as \code{all_lists}, containing the \code{i}th component of each element in \code{all_lists}. If any components of \code{all_lists} have fewer than \code{i} elements, \code{get_index_list()} will throw an error.
}
\description{
Extracts the \code{i}th component of each element in \code{all_lists}. This function is nearly identical to \code{get_index_vec()}, but instead correctly deals with elements themselves being lists.
}
\examples{
A = list(1,2,3)
B = list("a", "b")
all_lists = list(A, B)
get_index_list(all_lists, 2)
}
\keyword{internal}
