% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_index_vec}
\alias{get_index_vec}
\title{Extract the \code{i}th element of each vector in \code{all_vecs}}
\usage{
get_index_vec(all_vecs, i)
}
\arguments{
\item{all_vecs}{A list of vectors. Throws an error if any contain fewer than \code{i} entries.}

\item{i}{Index to extract.}
}
\value{
A vector of the same length as \code{all_vecs}, containing the \code{i}th component of each element in \code{all_vecs}. If any components of \code{all_vecs} have fewer than \code{i} elements, \code{get_index()} will throw an error.
}
\description{
Extracts the \code{i}th component of each element in \code{all_vecs}. This function is nearly identical to \code{get_index_list()}, but requires that no element of \code{all_vecs} be a list.
}
\examples{
A = c(1,2,3)
B = c("a", "b")
C = 1:1000
all_vecs = list(A, B, C)
get_index_vec(all_vecs, 2)
}
\keyword{internal}
