% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Validation_Data.R
\name{get_lin_preds}
\alias{get_lin_preds}
\title{Compute the linear predictors in each group}
\usage{
get_lin_preds(data_by_K, beta, u_by_K)
}
\arguments{
\item{data_by_K}{A list of length K, where each element is a list with first component the fixed effects design matrix and second component the random effects design matrix.}

\item{beta}{A vector of fixed effects.}

\item{u_by_K}{A list of length K, where each element is a vector of random effects for group k.}
}
\value{
A list of length K. Each element is a vector of length \code{n} containing the linear predictor for each observation in group \code{k}.
}
\description{
Compute the linear predictors in each group
}
\examples{
K=3
n=20
p=3
q=2

data_by_K = make_dataset(K, n, p, q)
beta = c(1,2,3)
u_by_K = list(c(1,1), c(2,2), c(3,3))
get_lin_preds(data_by_K, beta, u_by_K)
}
