% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_med_effs.R
\name{get_med_effs}
\alias{get_med_effs}
\title{Compute mediation effects based on regression parameters, respecting the variable types of Y and M}
\usage{
get_med_effs(X_in_Y, M_in_Y, X_in_M, Y_binary = TRUE, M_binary = TRUE)
}
\arguments{
\item{X_in_Y}{Coefficient of X in the regression model for Y.}

\item{M_in_Y}{Coefficient of M in the regression model for Y.}

\item{X_in_M}{Coefficient of X in the regression model for M.}

\item{Y_binary}{Is the outcome variable binary (TRUE) or continuous (FALSE)}

\item{M_binary}{Is the mediator variable binary (TRUE) or continuous (FALSE)}
}
\value{
A vector of mediation effects on odds-ratio scale. Order is direct effect, indirect effect, total effect.
}
\description{
Compute mediation effects based on regression parameters, respecting the variable types of Y and M
}
\examples{
# Binary outcome, binary mediator
X_in_Y <- 1
M_in_Y <- 1
X_in_M <- 1
get_med_effs(X_in_Y, M_in_Y, X_in_M)
}
