% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_med_effs.R
\name{get_med_effs_binY_binM}
\alias{get_med_effs_binY_binM}
\title{Compute mediation effects with binary response and binary mediator}
\usage{
get_med_effs_binY_binM(X_in_Y, M_in_Y, X_in_M)
}
\arguments{
\item{X_in_Y}{Coefficient of X in the regression model for Y.}

\item{M_in_Y}{Coefficient of M in the regression model for Y.}

\item{X_in_M}{Coefficient of X in the regression model for M.}
}
\value{
A vector of mediation effects on odds-ratio scale. Order is direct effect, indirect effect, total effect.
}
\description{
Compute mediation effects with binary response and binary mediator
}
\examples{
X_in_Y <- 1
M_in_Y <- 1
X_in_M <- 1
get_med_effs_binY_binM(X_in_Y, M_in_Y, X_in_M)
}
