% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lin_pred_contrib}
\alias{lin_pred_contrib}
\title{Compute the contribution to the linear predictor due to the provided dataset and coefficient vector}
\usage{
lin_pred_contrib(data, beta, add_intercept = TRUE)
}
\arguments{
\item{data}{A dataset. Can be a data frame or a matrix (former is converted to latter internally)}

\item{beta}{A vector of coefficients. Length must either match number of columns of \code{data} or be 1 less, if \code{add_intercept} is \code{FALSE} or \code{TRUE} respectively. In the latter case, intercept should be the first element of beta.}

\item{add_intercept}{Should \code{data} be augmented with a column of \code{1}s to represent an intercept? If so, this column is added to the left of \code{data}.}
}
\value{
A vector of linear predictors with length equal to the number of rows in \code{data}.
}
\description{
Compute the contribution to the linear predictor due to the provided dataset and coefficient vector
}
\examples{
data = data.frame(X1 = c(1,0), X2 = c(0,1))

# No intercept
beta1 = c(1,2)
lin_pred_contrib(data, beta1, add_intercept = FALSE)

# With intercept
beta2 = c(1,2,3)
lin_pred_contrib(data, beta2, add_intercept = TRUE)
}
\keyword{internal}
