% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_2_tibble}
\alias{list_2_tibble}
\title{Construct a tibble by \code{rbind}-ing elements of a list. Add/Create group labels.}
\usage{
list_2_tibble(X_list, group_labels = NULL)
}
\arguments{
\item{X_list}{A list of objects able to be passed recursively to \code{rbind}. In particular, must all have the same number of columns.}

\item{group_labels}{Optionally, a vector of labels for each element of \code{X_list}. Must be either the same length as \code{X_list} or equal to the total number of rows among elements in \code{X_list}. If \code{NULL}, labels are \code{G1}, \code{G2},...}
}
\value{
A tibble containing all elements of \code{X_list} with an extra column for labels.
}
\description{
Construct a tibble by \code{rbind}-ing elements of a list. Add/Create group labels.
}
\examples{
a_matrix = matrix(c(1,0,0,1), nrow = 2)
another_matrix = matrix(c(0,1,1,0), nrow = 2)
some_matrices = list(a_matrix, another_matrix)

list_2_tibble(some_matrices)

matrices_names = c("Alice", "Bob")
list_2_tibble(some_matrices, matrices_names)
}
