% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_2_tibble_unnamed}
\alias{list_2_tibble_unnamed}
\title{Stack elements of \code{X_list} and convert to a tibble}
\usage{
list_2_tibble_unnamed(X_list)
}
\arguments{
\item{X_list}{A list of objects able to be passed recursively to \code{rbind}. In particular, must all have the same number of columns.}
}
\value{
A single tibble containing the elements of \code{X_list}.
}
\description{
Stack elements of \code{X_list} and convert to a tibble
}
\examples{
a_matrix = matrix(c(1,0,0,1), nrow = 2)
another_matrix = matrix(c(0,1,1,0), nrow = 2)
some_matrices = list(a_matrix, another_matrix)

list_2_tibble_unnamed(some_matrices)
}
\keyword{internal}
