% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Validation_Data.R
\name{make_dataset}
\alias{make_dataset}
\title{Construct a full dataset with K groups, each containing fixed and random effects design matrices}
\usage{
make_dataset(K, n, p, q)
}
\arguments{
\item{K}{Number of groups.}

\item{n}{Number of observations per group.}

\item{p}{Number of covariates with fixed effects.}

\item{q}{Number of covariates with random effects.}
}
\value{
A list of length K. Each element is a list of length 2. The first element is an n-by-p matrix of fixed effects covariates. The second element is an n-by-q matrix of random effects covariates. Both matrices are generated by \code{make_some_data()}.
}
\description{
Construct a full dataset with K groups, each containing fixed and random effects design matrices
}
\examples{
K=3
n=20
p=3
q=2

data_by_K = make_dataset(K, n, p, q)
}
