% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Mixed_Effects.R
\name{make_MEs}
\alias{make_MEs}
\title{Construct mixed effects from given fixed and random effects}
\usage{
make_MEs(fix_effs, ran_effs_by_K, mix_eff_ind_pairs)
}
\arguments{
\item{fix_effs}{Vector of fixed effects}

\item{ran_effs_by_K}{List containing one vector of random effects for each group}

\item{mix_eff_ind_pairs}{List of pairs of indices, (i,j), indicating which fixed (i) and random (j) effects have the same covariate. Note: Don't forget about the intercept if necessary}
}
\value{
A list containing one vector of all regression coefficients for each group. Fixed and mixed effects come first, followed by any random-only effects.
}
\description{
Construct mixed effects from given fixed and random effects
}
\examples{
fix_effs = c(1,2,3)
K = 2
ran_effs_by_K = list(1, 2)
mix_eff_ind_pairs = list(c(2,1))
make_MEs(fix_effs, ran_effs_by_K, mix_eff_ind_pairs)
}
