% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Mixed_Effects.R
\name{make_REs}
\alias{make_REs}
\title{Generate random effects from given standard deviation matrix}
\usage{
make_REs(K, SD_mat)
}
\arguments{
\item{K}{Number of groups}

\item{SD_mat}{q-by-q standard deviation matrix of random effects (i.e. Cholesky factor of estimated random effects covariance matrix), where q is the number of covariates with random effects}
}
\value{
A list of random effect vectors. One element per group. Number of random effects to generate is inferred from the size of SD_mat
}
\description{
Generate random effects from given standard deviation matrix
}
\examples{
K=2
# Number of covariates with random effects in each group
q = 4
# SD matrix of random effects (i.e. Cholesky factor of its covariance matrix)
SD_mat = matrix(stats::rnorm(q^2), nrow = q, ncol = q)
all_REs = make_REs(K, SD_mat)
}
