% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Validation_Data.R
\name{make_some_data}
\alias{make_some_data}
\title{Construct an n-by-p matrix with standard normal elements}
\usage{
make_some_data(n, p)
}
\arguments{
\item{n}{Number of rows.}

\item{p}{Number of columns.}
}
\value{
An n-by-p matrix.
}
\description{
Construct an n-by-p matrix with standard normal elements
}
\examples{
n=20
p=3
make_some_data(n, p)
}
