% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_effs_from_lme4.R
\name{med_effs_from_lme4}
\alias{med_effs_from_lme4}
\title{Compute mediation effects from two regression models}
\usage{
med_effs_from_lme4(mod_Y, mod_M)
}
\arguments{
\item{mod_Y}{A regression model for the outcome variable, Y.}

\item{mod_M}{A regression model for the mediator variable, M.}
}
\value{
A matrix of mediation effects. Includes both fixed effects and group-specific mixed effects.
}
\description{
Compute mediation effects from two regression models
}
\examples{
mod_Y <- 1
mod_M <- 1
med_effs_from_lme4(mod_Y, mod_M)
}
