% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMediation-package.R
\docType{package}
\name{ModelMediation-package}
\alias{ModelMediation}
\alias{ModelMediation-package}
\title{ModelMediation: Perform model-based mediation analysis using a wide range of model types}
\description{
Perform model-based mediation analysis. Facilitates binary or continuous response and mediator, single or multiple levels and the option to include confounding variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wruth1/ModelMediation}
  \item Report bugs at \url{https://github.com/wruth1/ModelMediation/issues}
}

}
\author{
\strong{Maintainer}: William Ruth \email{William.Ruth@umontreal.ca} (\href{https://orcid.org/0000-0002-4975-1572}{ORCID})

}
\keyword{internal}
