% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_med_effs.R
\name{mult_med_effs}
\alias{mult_med_effs}
\alias{get_med_effs_DF}
\alias{get_med_effs_lme4}
\title{Compute mediation effects from regression coefficients}
\usage{
get_med_effs_DF(all_reg_coeffs)

get_med_effs_lme4(mod_Y, mod_M)
}
\arguments{
\item{all_reg_coeffs}{A data frame containing regression coefficients for mediation and, optionally, one or more other identifying variables.}

\item{mod_Y, mod_M}{Regression models for predicting the outcome and mediators respectively}
}
\value{
A data frame of mediation effects. Optionally also contains other identifying variables (e.g. group labels)
}
\description{
\code{get_med_effs_DF()} takes as input a data frame of regression coefficients. Specifically, this data frame must contain variables called \code{X_in_Y}, \code{M_in_Y} and \code{X_in_M}. Optionally, this data frame may contain other identifying variables, such as group label or bootstrap replicate number. These identifying variables are included in the output.

\code{get_med_effs_lme4()} takes as input two regression models fit using the \code{lme4} package, one for predicting the outcome variable, \code{Y}, the other for predicting the mediator variable, \code{M}. Relevant coefficients and group labels are extracted from these model objects, and are then passed to \code{get_med_effs_DF()} to compute mediation effects. Final output contains a column of group labels and no other identifying variables.
}
\examples{
n = 20
K = 3
all_reg_pars = make_all_reg_pars()
data = make_validation_data(n, K, all_reg_pars)

mod_Y = fit_mod_Y(data)
mod_M = fit_mod_M(data)

all_reg_coeffs = reg_coeffs_for_mediation(mod_Y, mod_M)
get_med_effs_DF(all_reg_coeffs)

# get_med_effs_lme4 combines the previous two steps
get_med_effs_lme4(mod_Y, mod_M)
}
