% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generic_Bootstrap.R
\name{one_bootstrap_sample}
\alias{one_bootstrap_sample}
\title{Generate a single bootstrap sample}
\usage{
one_bootstrap_sample(
  data = NULL,
  mod_Y = NULL,
  mod_M = NULL,
  parametric = TRUE,
  .careful = TRUE
)
}
\arguments{
\item{data}{Original dataset. Optional for parametric (as long as \code{mod_Y} and \code{mod_M} are supplied), required for non-parametric, .}

\item{mod_Y}{Regression model for predicting \code{Y}. Recommended but optional for parametric (as long as \code{data} is supplied), not used for non-parametric.}

\item{mod_M}{Regression model for predicting \code{M}. Recommended but optional for parametric (as long as \code{data} is supplied), not used for non-parametric.}

\item{parametric}{Indicates whether to perform parametric or non-parametric bootstrap.}

\item{.careful}{Should input be checked for whether enough information has been provided for the requested flavour of bootstrap? Recommended setting is \code{TRUE} if running interactively and \code{FALSE} if running inside another function which has already checked.}
}
\value{
A single bootstrap sample, formatted as a data.frame.
}
\description{
Generate a single bootstrap sample
}
\examples{
data = 0
mod_Y = fit_mod_Y(data)
mod_M = fit_mod_M(data)

one_bootstrap_sample(mod_Y = mod_Y, mod_M = mod_M, parametric = TRUE, .careful = TRUE)

one_bootstrap_sample(data = data, parametric = FALSE, .careful = TRUE)
}
