% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Samplers.R
\name{one_non_parametric_sample}
\alias{one_non_parametric_sample}
\title{Construct a non-parametric bootstrap sample from the provided data, conditional on group structure}
\usage{
one_non_parametric_sample(data)
}
\arguments{
\item{data}{A data frame. Must contain a variable called \code{group}.}
}
\value{
A new data frame generated by subsampling independently within each group.
}
\description{
Construct a non-parametric bootstrap sample from the provided data, conditional on group structure
}
\examples{
n = 20
K = 3
all_reg_pars = make_all_reg_pars()
data = make_validation_data(n, K, all_reg_pars)

one_non_parametric_sample(data)
}
