% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_samp_2_coeffs.R
\name{reg_coeffs_for_mediation}
\alias{reg_coeffs_for_mediation}
\title{Extract regression coefficients which are relevant for mediation analysis}
\usage{
reg_coeffs_for_mediation(mod_Y, mod_M)
}
\arguments{
\item{mod_Y, mod_M}{Regression models for predicting the outcome and mediators respectively}
}
\value{
A data frame containing regression coefficients for mediation analysis, as well as group labels
}
\description{
Extract regression coefficients which are relevant for mediation analysis
}
\examples{
n = 20
K = 3
all_reg_pars = make_all_reg_pars()
data = make_validation_data(n, K, all_reg_pars)

mod_Y = fit_mod_Y(data)
mod_M = fit_mod_M(data)

reg_coeffs_for_mediation(mod_Y, mod_M)
}
