% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_analysis}
\alias{run_analysis}
\title{Generate bootstrap samples, compute estimates, and construct CIs}
\usage{
run_analysis(data, B, .parallel = FALSE, .verbose = FALSE)
}
\arguments{
\item{data}{Observed dataset.}

\item{B}{Number of bootstrap replicates.}

\item{.parallel}{Should bootstrapping be performed in parallel?}
}
\value{
A data frame containing confidence intervals for each mediation effect, in each group, for all flavours of bootstrap and bootstrap CI.
}
\description{
Generate bootstrap samples, compute estimates, and construct CIs
}
\examples{
1+1
}
