% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_analysis_one_bootstrap}
\alias{run_analysis_one_bootstrap}
\title{Generate and return a single bootstrap sample of each flavour. Data must be formatted as a formal data frame.}
\usage{
run_analysis_one_bootstrap(
  data_formal,
  mod_Y = NULL,
  mod_M = NULL,
  .parallel = FALSE,
  .verbose = FALSE
)
}
\arguments{
\item{data_formal}{Observed dataset, structured as a formal data frame}

\item{mod_Y, mod_M}{Pre-fit models for the outcome and mediator, respectively. If NULL, these models will be fit inside the function.}

\item{.parallel}{Should bootstrapping be performed in parallel?}

\item{.verbose}{Should progress bars be produced for the two bootstrap analyses?}
}
\value{
A list containing one bootstrap sample of each flavour (parametric, semiparametric and nonparametric, in that order). Each sample contains the three mediation effects, both in aggregate and for each group.
}
\description{
Generate and return a single bootstrap sample of each flavour. Data must be formatted as a formal data frame.
}
\examples{
1+1
}
