% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap_Logic.R
\name{run_bootstrap}
\alias{run_bootstrap}
\title{Run a complete bootstrap analysis with \code{B} bootstrap samples}
\usage{
run_bootstrap(
  B,
  data = NULL,
  mod_Y = NULL,
  mod_M = NULL,
  parametric = TRUE,
  .parallel = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{B}{Number of bootstrap samples.}

\item{data}{Original dataset. Optional for parametric (as long as \code{mod_Y} and \code{mod_M} are supplied), required for non-parametric, .}

\item{mod_Y}{Regression model for predicting \code{Y}. Recommended but optional for parametric (as long as \code{data} is supplied), not used for non-parametric.}

\item{mod_M}{Regression model for predicting \code{M}. Recommended but optional for parametric (as long as \code{data} is supplied), not used for non-parametric.}

\item{parametric}{Indicates whether to perform parametric or non-parametric bootstrap.}

\item{.parallel}{Should the \code{B} analyses be performed in parallel?}

\item{.verbose}{Should a progress bar be printed?}
}
\value{
Standard output from a bootstrap analysis. Ready to be used for constructing confidence intervals.
}
\description{
Run a complete bootstrap analysis with \code{B} bootstrap samples
}
\examples{
B = 2
n = 20
K = 3
all_reg_pars = make_all_reg_pars()
data = make_validation_data(n, K, all_reg_pars)

mod_Y = fit_mod_Y(data)
mod_M = fit_mod_M(data)

run_bootstrap(B, mod_Y = mod_Y, mod_M = mod_M, parametric = TRUE, .parallel = FALSE)

run_bootstrap(B, data = data, parametric = FALSE, .parallel = FALSE)
}
