% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Make_Validation_Data.R
\name{make_all_reg_pars}
\alias{make_all_reg_pars}
\title{Store all regression parameters in a single list}
\usage{
make_all_reg_pars(
  beta_Y = rep(1, times = 4),
  Gamma_Y = matrix(rep(1, times = 9), nrow = 3),
  beta_M = rep(1, times = 3),
  Gamma_M = matrix(rep(1, times = 4), nrow = 2)
)
}
\arguments{
\item{beta_Y}{Vector of length 4 containing fixed-effects for the outcome model. Order of variables is: Intercept, X, M, C.}

\item{Gamma_Y}{Covariance matrix of size 3x3 containing random effects for the outcome model. Order of variables is: Intercept, X, M.}

\item{beta_M}{Vector of length 3 containing fixed-effects for the outcome model. Order is: Intercept, X, C.}

\item{Gamma_M}{Covariance matrix of size 2x2 containing random effects for the outcome model. Order of variables is: Intercept, X.}
}
\value{
A named list containing fixed and random effects parameters for both the outcome and mediator regression models. Names in list match names of arguments to this function.
}
\description{
Store all regression parameters in a single list
}
\examples{
make_all_reg_pars()
}
